requirements.txt
streamlit
openai
PyMuPDF
pandas
python-dateutil
altair
pytesseract
Pillow